Fix MSYS2 build: min/max, updater unicode, debug
Resolve MSYS2/C++20 and MinGW build regressions: export R_DrawArrowCap (remove static), include <algorithm> and import std::min/std::max in client header, and guard CG_UI_Com_DPrintf for non-debug builds. Make min() call sites explicit via size_t/ALint casts in cgame.cpp, screen.cpp, and sound/al.cpp, and fix netgraph clamp to use unsigned. Add GCC-only -municode to the worr_updater target link and compile args in meson.build and guard UNICODE/_UNICODE defines in worr_updater.c to avoid redefinition warnings. Add documentation of the MSYS2 nightly recovery and reference the fix in the feature roadmap.

+++ src/client/screen.cpp
@@ -341,7 +341,7 @@ void SCR_DrawStringMultiStretch(int x, int y, int scale, int flags, size_t maxle
             break;
         }
 
-        len = min(p - s, maxlen);
+        len = min(static_cast<size_t>(p - s), maxlen);
         last_x = SCR_DrawStringStretch(x, y, scale, flags, len, s, color, font);
         last_y = y;
         maxlen -= len;
@@ -375,7 +375,7 @@ void SCR_DrawKStringMultiStretch(int x, int y, int scale, int flags, size_t maxl
             break;
         }
 
-        int len = min(p - s, maxlen);
+        size_t len = min(static_cast<size_t>(p - s), maxlen);
         last_x = SCR_DrawKStringStretch(x, y, scale, flags, len, s, color, kfont);
         last_y = y;
         maxlen -= len;
@@ -869,7 +869,7 @@ void SCR_AddNetgraph(void)
         color = 0xd0;
     }
 
-    SCR_DebugGraph(min(ping, 30), color);
+    SCR_DebugGraph(min(ping, 30u), color);
 }
 
 #define GRAPH_SAMPLES   4096
